Move BranchDeserializer to its own file
