+++ src/main/java/com/spotify/github/v3/repos/Branch.java
@@ -21,20 +21,11 @@
 package com.spotify.github.v3.repos;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.core.JsonParser;
-import com.fasterxml.jackson.databind.DeserializationContext;
-import com.fasterxml.jackson.databind.JsonDeserializer;
-import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import com.spotify.github.GithubStyle;
-import com.spotify.github.v3.git.ImmutableShaLink;
 import com.spotify.github.v3.git.ShaLink;
-import java.io.IOException;
 import java.net.URI;
-import java.net.URISyntaxException;
-import java.net.URLEncoder;
-import java.nio.charset.StandardCharsets;
 import java.util.Optional;
 import javax.annotation.Nullable;
 import org.immutables.value.Value;
@@ -62,45 +53,3 @@ public interface Branch {
   Optional<URI> protectionUrl();
 }
 
-class BranchDeserializer extends JsonDeserializer<ImmutableBranch> {
-
-  private URI fixInvalidGithubUrl(final String invalidUrl) throws IOException {
-    // There's a bug in github where it gives you back non-url-encoded characters
-    // in the protection_url field. For example if your branch has a single "%" in its name.
-    // As of this writing, the protection URL looks something like this
-    // https://github-server.tld/api/v3/repos/owner/repo-name/branches/branch-name/protection
-    final String[] schemaAndPath = invalidUrl.split("//");
-    String[] pathParts = schemaAndPath[1].split("/");
-    for (int i = 0; i < pathParts.length; i++) {
-      pathParts[i] = URLEncoder.encode(pathParts[i], StandardCharsets.UTF_8);
-    }
-    String fixedUrlString = schemaAndPath[0] + "//" + String.join("/", pathParts);
-    return URI.create(fixedUrlString);
-  }
-
-  @Override
-  public ImmutableBranch deserialize(final JsonParser jsonParser, final DeserializationContext deserializationContext)
-      throws IOException {
-    JsonNode node = jsonParser.getCodec().readTree(jsonParser);
-    URI protectionUrl;
-    var immutableBranchBuilder = ImmutableBranch.builder();
-    if (node.has("protected")) {
-      immutableBranchBuilder.isProtected(node.get("protected").asBoolean());
-      String protectionUrlString = node.get("protection_url").asText();
-      try {
-        protectionUrl = new URI(protectionUrlString);
-      } catch (URISyntaxException e) {
-        protectionUrl = fixInvalidGithubUrl(protectionUrlString);
-      }
-      immutableBranchBuilder.protectionUrl(protectionUrl);
-    }
-    ImmutableShaLink shaLink = ImmutableShaLink.builder()
-        .sha(node.get("commit").get("sha").asText())
-        .url(URI.create(node.at("/commit/url").asText()))
-        .build();
-    return immutableBranchBuilder
-        .name(node.get("name").textValue())
-        .commit(shaLink)
-        .build();
-  }
-}
